bluetooth: implement writeValueWith[out]Response()
This implements the Web Bluetooth BluetoothRemoteGATTCharacteristic
writeValueWithResponse() and writeValueWithoutResponse() methods.

https://webbluetoothcg.github.io/web-bluetooth/#writecharacteristicvalue

The writeValue() method now prefers write with response if a remote
characteristic has both WRITE and WRITE_WITHOUT_RESPONSE properties.
This is a potentially breaking change on Android devices as the OS
default there seems to be to write without response. All other platforms
already defaulted to write with response when both properties are
present.

All implemented platforms now check the requested write type against
the reported properties before sending a message and throw
ERROR_NOT_SUPPORTED if the requested write is not possible. This is
done to match the Web Bluetooth specification but could be a breaking
change since some platforms previously threw ERROR_NOT_PERMITTED
instead.

Android, bluez, mac and winrt have been implemented and tested on real
hardware. (Chrome)cast and win32 have been implemented but not tested.

Also fixed spelling of "already" while we are touching this code.

Bug: 672648
Change-Id: I0d8032fded0028eaa78a55afe3336f253c85c590